Add Homography Tracker API (#1389)
* rebase commit

* init

* added test data for loftr and image registrator

* real ransac tests

* save

* works

* lint

* üpdated example

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* docs

* docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* stupid CI OOM

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* lint

* fix dtype in tests

* fix test dtype

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* no-grad for homotest

* fix

* fix

* fix random seed for ransac test

* some formatting

* codespell

* fix signature defaults

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add fixture with data

* remove files

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix deepsource

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* deepsource

* fix super

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* identation fix

* fix positional

* fix mypy

* refactor some tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* remove doctest

* fix ransac tests

* fix similarity tests

* fix sha path

* fix loftr precision

* xfail some ransac tests

+++ kornia/feature/loftr/utils/coarse_matching.py
@@ -262,8 +262,8 @@ class CoarseMatching(nn.Module):
         coarse_matches.update({
             'gt_mask': mconf == 0,
             'm_bids': b_ids[mconf != 0],  # mconf == 0 => gt matches
-            'mkpts0_c': mkpts0_c[mconf != 0],
-            'mkpts1_c': mkpts1_c[mconf != 0],
+            'mkpts0_c': mkpts0_c[mconf != 0].to(dtype=conf_matrix.dtype),
+            'mkpts1_c': mkpts1_c[mconf != 0].to(dtype=conf_matrix.dtype),
             'mconf': mconf[mconf != 0]
         })
